+++ src/js/cookies.js
@@ -9,7 +9,9 @@ Document.prototype.$cookies = {
   },
 
   remove: function(key) {
+    var value = this.get(key);
     document.cookie = key + "=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/";
+    return value;
   },
 
   clear: function() {